JAV-438 fix pr comment
